requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
Django==4.0
django-environ==0.9.0
django-filter==22.1
django-leaflet==0.28.3
djangorestframework==3.13.0
djangorestframework-gis==1.0
environs==9.5.0
flake8==5.0.4
install==1.3.5
marshmallow==3.17.0
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
